Fix : api usings & moved mapper
