+++ api/src/FinancialHub/FinancialHub.WebApi/FinancialHub.WebApi.csproj
@@ -27,6 +27,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\FinancialHub.Infra\FinancialHub.Infra.Data.csproj" />
+    <ProjectReference Include="..\FinancialHub.Services\FinancialHub.Services.csproj" />
   </ItemGroup>
 
 </Project>
